Move DeadCodeRemover to iter_visitor.cpp
