+++ csrc/optimization/remove_empty.cpp
@@ -8,6 +8,7 @@
 #include <optimization/remove_empty.h>
 
 #include <ir/utils.h>
+#include <iter_visitor.h>
 #include <ops/alias.h>
 #include <ops/arith.h>
 
@@ -19,165 +20,6 @@
 namespace nvfuser::optimization {
 
 namespace {
-
-//! This is a generic traversal class that is used to modify a Fusion graph by
-//! replacing TensorViews so that their definitions can be altered to simplify
-//! computation or remove dead code. This differs from OptOutMutator, which is
-//! built for mutating TensorViews in a graph, and does not easily handle
-//! modifying TensorView definitions and Expr Fusion inputs during traversal.
-//!
-//! We use unordered_set called live_statements_, which is initialized as the
-//! Exprs in traversal_exprs_ as well as their inputs and their outputs with
-//! live uses. Marking a Statement as dead removes it from live_statements_,
-//! and replacing a Val inserts the Val and its definition, recursively. Since
-//! we traverse backwards, and we handle all active Expr outputs, this ensures
-//! that it is safe to removing an Expr will not result in erasing definitions
-//! of active Expr outputs.
-//!
-//! Derived classes should override handle() and make use of replaceTV(),
-//! markDeadAndMaybeRemove(), and allUsesDead(). Note that if replacements are
-//! made using replaceTV(old_tv, new_tv), then neither new_tv or any new
-//! Statements produced in creating it will be traversed by this class.
-class DeadCodeRemover : BackwardVisitor {
- public:
-  DeadCodeRemover(Fusion* fusion) : BackwardVisitor(false), fusion_(fusion) {}
-
-  void run() {
-    // First we build a set of all live Statements so that we can detect dead
-    // branches.
-    for (auto stmt : StmtSort::getStmts(fusion_, fusion_->outputs())) {
-      markLive(stmt);
-    }
-
-    // Note that StmtSort::getStmts() is also run in traverseTo. In the future,
-    // we could potentially refactor this so that derived classes from
-    // BackwardVisitor can make use of that traversal instead of repeating it.
-    traverseTo(fusion_, fusion_->outputs());
-  }
-
-  Fusion* fusion() const {
-    return fusion_;
-  }
-
- protected:
-  using BackwardVisitor::handle;
-
-  void handle(Statement* stmt) final {
-    if (!isLive(stmt)) {
-      // We check whether stmt is dead before we dereference it, since it may
-      // have been removed from the Fusion.
-      return;
-    }
-    BackwardVisitor::handle(stmt);
-  }
-
-  void handle(Expr* expr) final {
-    if (allOutputsDead(expr)) {
-      markDead(expr);
-      fusion_->removeExpr(expr);
-    } else {
-      BackwardVisitor::handle(expr);
-    }
-  }
-
-  //! Replaces a TensorView in outputs, and in all uses. If old_tv is a Fusion
-  //! input, we do not replace it. After replacement, unless it is a Fusion
-  //! input, we remove it from the fusion and set the original pointer to zero
-  //! (hence why old_tv is passed by reference).
-  void replaceTV(TensorView*& old_tv, TensorView* new_tv) {
-    if (old_tv->isFusionOutput()) {
-      fusion_->replaceOutput(old_tv, new_tv);
-    }
-    for (auto use : old_tv->uses()) {
-      ir_utils::replaceValInExpr(use, old_tv, new_tv);
-    }
-    markLiveRecursive(new_tv);
-    markDeadAndMaybeRemove(old_tv);
-  }
-
-  //! Guard removeVal with calls to markDead so that we always detect removed
-  //! Vals and Exprs before derefencing them.
-  //!
-  //! Note that we only remove val if all of its definition's outputs are marked
-  //! dead.
-  void markDeadAndMaybeRemove(Val* val) {
-    markDead(val);
-    if (val->definition()) {
-      if (allOutputsDead(val->definition())) {
-        // If all other outputs are dead, it's safe to remove the definition as
-        // well as all its outputs
-        markDead(val->definition());
-        const auto outputs = val->definition()->outputs();
-        for (auto outp : outputs) {
-          fusion_->removeVal(outp);
-        }
-      }
-    } else {
-      TORCH_INTERNAL_ASSERT(
-          !val->isFusionInput(), "Refusing to remove Fusion input");
-      fusion_->removeVal(val);
-    }
-  }
-
-  //! Find whether a statement is live (i.e. not dead code)
-  //!
-  //! Inside BackwardVisitor::traverseTo, traversal_exprs_ is built, containing
-  //! all active expressions in the original graph.
-  bool isLive(Statement* stmt) {
-    return live_statements_.find(stmt) != live_statements_.end();
-  }
-
-  //! Check whether all outputs of an expression have been marked dead
-  bool allOutputsDead(Expr* expr) {
-    return std::all_of(
-        expr->outputs().begin(), expr->outputs().end(), [&](Val* outp) {
-          return !isLive(outp);
-        });
-  }
-
-  //! Check whether all uses have been marked dead
-  bool allUsesDead(Val* val) {
-    for (const auto use : val->uses()) {
-      if (isLive(use)) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  //! Mark a single Statement as being alive
-  void markLive(Statement* stmt) {
-    live_statements_.insert(stmt);
-  }
-
-  //! Ensure that a Statement and its upstream Statements are alive. If it is an
-  //! Expr, ensure all its inputs are alive. If it's a Val with a definition,
-  //! recursive to the definition.
-  void markLiveRecursive(Statement* stmt) {
-    if (isLive(stmt)) {
-      return;
-    }
-    markLive(stmt);
-    if (stmt->isVal() && stmt->asVal()->definition()) {
-      markLiveRecursive(stmt);
-    } else {
-      for (const auto inp : stmt->asExpr()->inputs()) {
-        markLiveRecursive(inp);
-      }
-    }
-  }
-
-  //! Mark a Statement* as dead so that we avoid dereferencing it later
-  void markDead(Statement* stmt) {
-    live_statements_.erase(stmt);
-  }
-
- private:
-  Fusion* fusion_;
-
-  //! Statements are marked dead by removing them from this set
-  std::unordered_set<Statement*> live_statements_;
-};
 
 //! Get a vector of the integer positions of constant zero extent axes in the
 //! input domain. This will typically be used like
